plagium_core/requirements.txt
antlr4-python3-runtime==4.13.1
flask-cors==4.0.0
pip-chill==1.0.3
python-dotenv==1.0.0
